1.8.1 Fix some documentation lines.

+++ src/main/java/com/bloodnbonesgaming/topography/world/generator/CellNoiseGenerator.java
@@ -68,7 +68,7 @@ public class CellNoiseGenerator implements IGenerator
     	this.frequency = frequency;
     }
     
-    @ScriptMethodDocumentation(args = "float", usage = "cutoff", notes = "Sets the noise cutoff for blocks to be placed. Higher values = more blocks. Default is -0.15.")
+    @ScriptMethodDocumentation(args = "double", usage = "cutoff", notes = "Sets the noise cutoff for blocks to be placed. Higher values = more blocks. Default is -0.15.")
 	public void setCutoff(final double cutoff)
     {
     	this.cutoff = cutoff;
